refactor: STORIF-335 - AccessKeyDrawer relpaced by OMC_AccessKeyDrawer.
