+++ packages/manager/src/features/ObjectStorage/AccessKeyLanding/utils.test.ts
@@ -4,7 +4,7 @@ import {
   hasLabelOrRegionsChanged,
 } from './utils';
 
-import type { DisplayedAccessKeyScope, FormState } from './OMC_AccessKeyDrawer';
+import type { DisplayedAccessKeyScope, FormState } from './AccessKeyDrawer';
 import type { ObjectStorageKey } from '@linode/api-v4/lib/object-storage';
 
 describe('generateUpdatePayload', () => {
