+++ packages/manager/src/features/ObjectStorage/AccessKeyLanding/utils.ts
@@ -1,6 +1,6 @@
 import { areArraysEqual, sortByString } from '@linode/utilities';
 
-import type { DisplayedAccessKeyScope, FormState } from './OMC_AccessKeyDrawer';
+import type { DisplayedAccessKeyScope, FormState } from './AccessKeyDrawer';
 import type { ObjectStorageKey } from '@linode/api-v4/lib/object-storage';
 
 type UpdatePayload =
